Final_Project/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pypdf2
python-multipart